ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
langchain==0.3.7
langchain-text-splitters==0.3.2
langchain-ollama==0.2.0
langchain-chroma==0.1.4
langgraph==0.2.39
langgraph-checkpoint-sqlite==2.0.2
pypdf==5.0.1
python-docx==1.1.2
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
python-json-logger==2.0.7
tenacity==9.0.0
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2
